requirements.txt
aenum==3.1.17
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
Flask==3.1.3
frozenlist==1.8.0
future==1.0.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gspread==6.2.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.22.0
MarkupSafe==3.0.3
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
yarl==1.23.0
